Move fontawesome.js into js/
