+++ src/js/fontawesome.js
